+++ CHANGELOG.md
@@ -5,6 +5,7 @@ All notable changes to the "vscode-debug-adapter-apache-camel" extension will be
 ## 0.10.0
 
 - Update default Camel version used for Camel JBang from 4.0.0 to 4.2.0
+- Provide build and start tasks completion inside `.vscode/tasks.json` for a Camel Quarkus native application
 
 ## 0.9.0